+++ server/src/http/controllers/jobs/jobs.controller.ts
@@ -29,7 +29,7 @@ import { IApiError } from "../../../common/utils/errorManager.js"
 import { ILbaItem } from "../../../services/lbaitem.shared.service.types.js"
 import { getCompanyFromSiret } from "../../../service/poleEmploi/bonnesBoites.js"
 import { addOffreDetailView, addOffreSearchView, getLbaJobById } from "../../../services/lbajob.service.js"
-import { getPeJobFromId } from "../../../service/poleEmploi/offresPoleEmploi.js"
+import { getPeJobFromId } from "../../../services/pejob.service.js"
 import { getJobsQuery } from "../../../service/poleEmploi/jobsAndCompanies.js"
 
 @Tags("Jobs")
